# docs/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Salam Həyətimiz API
  version: 1.2.0
  summary: REST API for the Salam Həyətimiz remote access-control platform.
  description: |
    # Overview

    This document is the **binding contract** between the Laravel backend and all clients
    (Flutter mobile app, Blade admin panel, and Kapital Bank webhooks). It is the source
    of truth referenced by `TECHNICAL_SPECIFICATION.md` §9 and aligns with `DATABASE_ARCHITECTURE.md` v1.0.

    The API is split into two surfaces:

    | Surface | Prefix | Audience | Auth |
    |---|---|---|---|
    | Mobile / Public | `/v1` | End users, device owners, technical mobile mode | Phone + OTP → JWT (`bearerAuth`) |
    | Admin | `/admin/v1` | Super admin, technical (back-office) | Email + password + TOTP → JWT (`adminBearerAuth`) |
    | Webhooks | `/v1/payments/callback` | Kapital Bank | HMAC signature + IP allowlist |

    All times are **UTC** in **RFC 3339** form. All money is in **minor units** (qəpik;
    1 AZN = 100). Currency is always **AZN**.

    # Authentication

    ## Mobile (`bearerAuth`)
    Phone-based OTP authentication. The flow:

    1. `POST /auth/otp/request` with phone — backend dispatches a 6-digit SMS OTP (TTL 120 s, max 5 attempts).
    2. `POST /auth/otp/verify` with phone + code + device fingerprint — backend issues a JWT access token (15 min) and an opaque refresh token (60 days).
    3. Send `Authorization: Bearer <access_token>` on subsequent calls.
...

...

...

...

...

...

    # Authorization

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []

# =============================================================================
# PATHS
# =============================================================================
paths:
...
  /devices:
    get:
      tags: [Devices]
      summary: List the current user's devices (owned + member)
      operationId: listMyDevices
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - in: query
          name: filter
          schema:
            type: string
            enum: [all, owned, member, suspended]
            default: all
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceListResponse' }
  /devices/{deviceId}:
    parameters: [{ $ref: '#/components/parameters/DeviceId' }]
    get:
      tags: [Devices]
      summary: Device detail
      operationId: getDevice
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /devices/{deviceId}/open:
    parameters: [{ $ref: '#/components/parameters/DeviceId' }]
    post:
      tags: [Commands]
      summary: Issue an open command
      operationId: openDevice
      description: |
        Idempotent. The `Idempotency-Key` is also persisted as `open_commands.idempotency_key`
        — replays return the cached command without dispatching again.

        Authorization: user must be on the device roster, subscription must be `active`,
        device must not be `disabled`.

      x-authorization: user; must be on roster with active subscription
      x-rate-limit: 12/user/min, 4/device/min
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                client_app_version: { type: string, maxLength: 20 }
                latitude:
                  type: number
                  format: double
                  minimum: -90
                  maximum: 90
                  description: |
                    *(GEOFENCE-1)* Caller's current latitude. Required only when the device's
                    `geofence_enabled` is true (else the server rejects with `location_required`).
                longitude:
                  type: number
                  format: double
                  minimum: -180
                  maximum: 180
                  description: Caller's current longitude (see `latitude`).
                accuracy:
                  type: number
                  format: double
                  minimum: 0
                  description: |
                    Horizontal accuracy of the fix in metres. Applied as a tolerance margin; a fix
                    whose accuracy exceeds the device radius is rejected as `location_imprecise`.
      responses:
        '202':
          description: Command queued; poll `/commands/{id}` or subscribe via WebSocket
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OpenCommandAccepted' }
        '403':
          description: |
            subscription_required / device_disabled / forbidden /
            location_required / outside_geofence / location_imprecise (GEOFENCE-1)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '429':
          description: Cooldown or global rate limit
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
          headers:
            Retry-After:
              schema: { type: integer }
              description: Seconds until next allowed attempt
        '502':
          description: Device unreachable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /devices/{deviceId}/commands:
    parameters: [{ $ref: '#/components/parameters/DeviceId' }]
    get:
      tags: [Commands]
      summary: Open-command history for the device (current user)
      operationId: listDeviceCommands
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - in: query
          name: since
          schema: { type: string, format: date-time }
        - in: query
          name: until
          schema: { type: string, format: date-time }
        - in: query
          name: state
          schema: { type: string, enum: [queued, dispatched, opened, failed, expired] }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OpenCommandListResponse' }
  /commands/{commandId}:
    parameters:
      - in: path
        name: commandId
        required: true
        schema: { type: integer, format: int64, minimum: 1 }
    get:
      tags: [Commands]
      summary: Get a single open command (terminal state polling)
      operationId: getCommand
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OpenCommand' }
        '404': { $ref: '#/components/responses/NotFound' }
  /commands/{commandId}/feedback:
    parameters:
      - in: path
        name: commandId
        required: true
        schema: { type: integer, format: int64, minimum: 1 }
    post:
      tags: [Commands]
      summary: User-reported actuation confirmation (v1.1, CRIT-06)
      operationId: submitOpenFeedback
      description: |
        Used after opens where `driver_confirms_actuation = false` to capture whether the
        gate actually moved. One feedback per (command, user); replays return the previous value.
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [gate_moved]
              properties:
                gate_moved: { type: boolean }
                comment: { type: string, maxLength: 255, nullable: true }
      responses:
        '201': { description: Feedback recorded }
        '200': { description: Replay; feedback already on record }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------------------
  # Roster (owner only)
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Mobile JWT issued by `/auth/otp/verify` or `/auth/refresh`.
        Algorithm RS256, 15-minute TTL. Subject is `user.{id}`.
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Admin JWT issued by `/admin/v1/auth/2fa/verify`. Disjoint identity space
        from `bearerAuth`. Algorithm RS256, 30-minute TTL.
    bankCallbackSignature:
      type: apiKey
      in: header
      name: X-Kapital-Signature
      description: |
        HMAC-SHA256 hex digest of the canonical request body using the shared
        secret negotiated with Kapital Bank. Verified before any state mutation.
    smsProviderSignature:
      type: apiKey
      in: header
      name: X-Sms-Signature
      description: HMAC signature from the inbound SMS provider.

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
    DeviceId:
      in: path
      name: deviceId
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
...
    Limit:
      in: query
      name: limit
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Cursor:
      in: query
      name: cursor
      required: false
      schema: { type: string, maxLength: 200 }
      description: Opaque pagination cursor returned in previous `page.next_cursor`.
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema: { type: string, minLength: 1, maxLength: 60 }
      description: |
        Optional. Same `(scope, actor, key)` returns the cached response for 24 h.
        Same key + different body → `409 idempotency_mismatch`.
    IdempotencyKeyRequired:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, minLength: 1, maxLength: 60 }
      description: |
        Required. Same `(scope, actor, key)` returns the cached response for 24 h.

  # ---------------------------------------------------------------------------
  # Reusable Headers
  # ---------------------------------------------------------------------------
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    IdempotencyConflict:
      description: Same Idempotency-Key with a different request body
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: subscription_required
            message_key:
              type: string
              example: errors.subscription_required
            message:
              type: string
              example: Aboneliyiniz başa çatıb.
            details:
              type: object
              additionalProperties: true
              nullable: true
            request_id:
              type: string
              example: 01JCGZ9F9XZ7P0X5MFK7Z7VYNZ
...
    PageInfo:
      type: object
      required: [has_more, limit]
      properties:
        next_cursor:
          type: string
          nullable: true
        has_more: { type: boolean }
        limit: { type: integer }
...
    Device:
      type: object
      required: [id, label, status, role, can_open]
      properties:
        id: { type: integer, format: int64 }
        label: { type: string, maxLength: 160, description: location_label or serial fallback }
        serial: { type: string, maxLength: 64 }
        status:
          type: string
          enum: [active, suspended, disabled, decommissioned]
        role: { type: string, enum: [owner, user] }
        can_open:
          type: boolean
          description: True if the caller is allowed to open right now (sub active, device active, not in cooldown)
        suspension_reason:
          type: string
          description: |
            *(v1.1, CRIT-02)* Per-caller derived reason. `owner_sub_expired_others_active`
            is shown to owners whose own sub has lapsed while invitees still have active
            subs — the device is open-capable for others but not for the owner.
          enum:
            - none
            - subscription_expired
            - owner_sub_expired_others_active
            - device_disabled
            - device_suspended
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        geofence_enabled:
          type: boolean
          description: |
            *(GEOFENCE-1)* When true the app MUST attach the caller's GPS
            (`latitude`/`longitude`, optional `accuracy`) to `POST /devices/{id}/open`, or the
            server rejects with `location_required`. The backend holds and enforces the radius.
        geofence_radius_m:
          type: integer
          nullable: true
          minimum: 1
          maximum: 65535
          description: Allowed distance from the device in metres (null when the geofence is off).
        last_online_at: { type: string, format: date-time, nullable: true }
        image_url: { type: string, maxLength: 2048, nullable: true, description: Barrier photo URL (null → app shows a placeholder) }
        address: { type: string, maxLength: 255, nullable: true, description: Full street address (null → app shows a fallback) }
    DeviceDetail:
      allOf:
        - { $ref: '#/components/schemas/Device' }
        - type: object
          properties:
            device_model:
              type: object
              properties:
                vendor: { type: string }
                model_code: { type: string }
            cooldown_seconds_remaining: { type: integer, nullable: true }
            subscription: { $ref: '#/components/schemas/SubscriptionBrief' }
...
    DeviceListResponse:
      type: object
      required: [data, page]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Device' }
        page: { $ref: '#/components/schemas/PageInfo' }
...
    OpenCommandAccepted:
      type: object
      required: [command_id, state, expected_completion_ms, driver_confirms_actuation]
      properties:
        command_id: { type: integer, format: int64 }
        state: { type: string, enum: [queued, dispatching] }
        expected_completion_ms:
          type: integer
          example: 5000
          description: |
            *(v1.1, HIGH-14)* Server-computed estimate based on the device's driver and a
            rolling p90 of recent successful opens. The UX SHOULD size its progress UI to
            this hint, not to a fixed constant. Defaults to a per-driver constant when the
            rolling window has < 10 samples.
        driver_confirms_actuation:
          type: boolean
          description: |
            *(v1.2, CRIT-06 resolved)* When `true`, the driver observes an actuation signal
            (e.g. Traccar output-state read-back / command ack, or a BLE script ack); the terminal
            state `opened` carries real evidence and the UX should render "Açıldı".
            When `false` (no actuation feedback channel), the terminal state means
            "dispatched"; the UX should render "Göndərildi" and may prompt for user
            confirmation via `submitOpenFeedback`.
        websocket_channel:
          type: string
          example: 'private-user.42'
    OpenCommand:
      type: object
      required: [id, device_id, state, requested_at]
      properties:
        id: { type: integer, format: int64 }
        device_id: { type: integer, format: int64 }
        state: { type: string, enum: [queued, dispatching, dispatched, opened, failed, expired] }
        failure_reason: { type: string, nullable: true }
        driver: { type: string, enum: [traccar, ble, sms] }
        requested_at: { type: string, format: date-time }
        dispatched_at: { type: string, format: date-time, nullable: true }
        completed_at: { type: string, format: date-time, nullable: true }
        latency_ms: { type: integer, nullable: true }
        attempts: { type: integer }
    OpenCommandListResponse:
      type: object
      required: [data, page]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/OpenCommand' }
        page: { $ref: '#/components/schemas/PageInfo' }

    # ----- Subscriptions ------------------------------------------------
    SubscriptionBrief:
      type: object
      properties:
        id: { type: integer, format: int64 }
        tier: { type: string, enum: [main, additional] }
        status: { type: string, enum: [pending_payment, active, expired, cancelled, refunded] }
        ends_at: { type: string, format: date-time, nullable: true }
        days_remaining: { type: integer, nullable: true }
        auto_renew: { type: boolean }
...